NCD/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
branca==0.7.2
cell-gears==0.0.2
chex==0.1.7
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contextlib2==21.6.0
datasets==2.20.0
dcor==0.6
decoupler==1.6.0
Deprecated==1.2.14
dill==0.3.8
dm-tree==0.1.8
docrep==0.3.2
et-xmlfile==1.1.0
etils==1.3.0
filelock==3.14.0
fiona==1.9.6
flax==0.7.2
folium==0.16.0
frozenlist==1.4.1
fsspec==2024.5.0
geopandas==0.13.2
hpack==4.0.0
huggingface-hub==0.23.4
inflect==7.2.1
jax==0.4.13
jaxlib==0.4.13
lightning-utilities==0.11.2
llvmlite==0.40.1
markdown-it-py==3.0.0
mdurl==0.1.2
missingno==0.5.2
ml-collections==0.1.1
ml-dtypes==0.2.0
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
mudata==0.2.3
multidict==6.0.5
multipledispatch==1.0.0
multiprocess==0.70.16
munkres==1.1.4
numpyro==0.12.1
omnipath==1.0.8
openpyxl==3.1.4
opt-einsum==3.3.0
optax==0.1.8
orbax==0.1.7
orbax-checkpoint==0.2.3
pandas==2.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydot==2.0.0
pyproj==3.5.0
pyro-api==0.1.2
pyro-ppl==1.9.1
pytorch-lightning==1.9.5
requests==2.32.3
rich==13.7.1
scgpt==0.2.1
scib==1.1.5
scikit-misc==0.2.0
scvi-tools==0.20.3
shapely==2.0.4
squarify==0.4.3
sympy==1.12.1
tbb==2021.10.0
tensorstore==0.1.45
timm==0.1.12
toolz==0.12.1
torch==2.1.2
torch_geometric==2.5.3
torchdata==0.7.1
torchmetrics==1.4.0.post0
torchtext==0.16.2
torchvision==0.17.2
tqdm==4.66.4
typeguard==4.3.0
typing_extensions==4.12.0
wrapt==1.16.0
xxhash==3.4.1
xyzservices==2024.4.0
yarl==1.9.4
